# Increase the performance of JSON marshaling