… exporter/splunkhecexporter/go.sum
… receiver/splunkhecreceiver/go.sum