@sethallen 2024-07-11
Just want to say this is an awesome improvement in performance @atoulme ! Thank you.